// src/enums/layouts.ts
export enum Layouts {
    // A standard Magic card with one face
    Normal = "normal",

    // A split-faced card
    Split = "split",

    // Cards that invert vertically with the flip keyword
    Flip = "flip",

    // Double-sided cards that transform
    Transform = "transform",

    // Cards with meld parts printed on the back
    Meld = "meld",

    // Cards with Level Up
    Leveler = "leveler",

    // Plane and Phenomenon-type cards
    Planar = "planar",

    // Scheme-type cards
    Scheme = "scheme",

    // Vanguard-type cards
    Vanguard = "vanguard",

    // Token cards
    Token = "token",

    // Tokens with another token printed on the back
    DoubleFacedToken = "double_faced_token",

    // Emblem cards
    Emblem = "emblem",

    // Cards with Augment
    Augment = "augment",

    // Host-type cards
    Host = "host"
}