// src/enums/powers.ts
export enum Powers {
    NegativeOne = "-1",
    Star = "*",
    Infinity = "∞",
    StarSquared = "*²",
    QuestionMark = "?",
    Zero = "0",
    PlusZero = "+0",
    OneHalf = ".5",
    PlusOne = "+1",
    One = "1",
    OnePlusStar = "1+*",
    OneAndOneHalf = "1.5",
    Two = "2",
    TwoPlusStar = "2+*",
    PlusTwo = "+2",
    TwoAndOneHalf = "2.5",
    PlusThree = "+3",
    Three = "3",
    ThreeAndOneHalf = "3.5",
    PlusFour = "+4",
    Four = "4",
    Five = "5",
    Six = "6",
    Seven = "7",
    Eight = "8",
    Nine = "9",
    Ten = "10",
    Eleven = "11",
    Twelve = "12",
    Thirteen = "13",
    Fifteen = "15",
    Twenty = "20",
    NinetyNine = "99"
}